requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32
jinja2>=3.1
gspread>=6.1
google-auth>=2.35
apscheduler>=3.10,<4.0
python-dotenv>=1.0
pydantic>=2.10
pytest>=8.3
